+++ src/main/kotlin/com/dh/baro/order/application/OrderFacade.kt
@@ -28,7 +28,7 @@ class OrderFacade(
 
     fun placeOrder(userId: Long, request: OrderCreateRequest): Order {
         userService.checkUserExists(userId)
-        val productList = productQueryService.getProductsExists(request.orderItems.map { orderItem -> orderItem.productId.toLong() })
+        val productList = productQueryService.getProductsExists(request.orderItems.map { orderItem -> orderItem.productId })
         val cmd = OrderCreateCommand.toCommand(userId, productList, request)
         val order = orderService.createOrder(cmd)
         return order
@@ -39,7 +39,7 @@ class OrderFacade(
         userService.checkUserExists(userId)
 
         val productList = productQueryService.getProductsExists(
-            request.orderItems.map { orderItem -> orderItem.productId.toLong() },
+            request.orderItems.map { orderItem -> orderItem.productId },
         )
 
         val cmd = OrderCreateCommand.toCommand(userId, productList, request)
